Support Obelisk HFS
The support is about as clean as other clients,
but the differences are so minor it shouldn't matter.

If in the future other kinds show up, we might have to make
async/Consumer and async/Producer encapsulate Client,
rather than subclass it.
That would easily allow swapping out of Clients.