Doc incompatibility of sync module with async enviro
One cannot use run_until_complete inside of a Future,
this is a limitation of the asyncio design.

Because our sync implementation is nothing but a run_until_complete
wrapper around the asynchronous,
it cannot be used in a Future environment.

This catches especially users of Jupyter by surprise, so we document it.